backend/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
psycopg2-binary
sqlalchemy
sqlalchemy-cockroachdb
python-dotenv